Fix typo in D3D12 sampler description conversion

+++ source/d3d12/d3d12_device.cpp
@@ -653,7 +653,7 @@ HRESULT STDMETHODCALLTYPE D3D12Device::CreateRootSignature(UINT nodeMask, const 
 
 				if (sampler_count != 0)
 				{
-					const uint32_t k = params.size();
+					const uint32_t k = static_cast<uint32_t>(params.size());
 					params.emplace_back();
 					ranges.emplace_back();
